user can reorganize, create, rename and delete his personnal folders
