+++ src/Components/PersonnalFolders/Assets.scss
@@ -2,8 +2,10 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  width: 100%;
   .folders_container {
     display: flex;
+    width: 100%;
     flex-direction: row;
     flex-wrap: wrap;
     justify-content: space-around;
@@ -19,16 +21,52 @@
       .asset_title {
         text-align: center;
         font-size: 20px;
+        width: 80%;
       }
     }
   }
 }
 
 .add_asset_container {
-  width: 100%;
+  position: absolute;
+  bottom: 40px;
+  right: 40px;
   display: flex;
   justify-content: center;
-  margin: 20px 0 20px 0;
+}
+
+.add_asset_modal {
+  background-color: #f8f9fa;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 400px;
+  transform: translate(-50%, -50%);
+  padding: 20px;
+  p {
+    margin-bottom: 20px;
+  }
+  div {
+    height: 40px;
+    width: 100%;
+    margin-bottom: 20px;
+  }
+  input {
+    width: 100%;
+  }
+  .add_asset_modal_action_bar {
+    display: flex;
+    justify-content: flex-end;
+    margin-bottom: 0;
+    button:first-of-type {
+      margin-right: 20px;
+    }
+  }
+}
+
+.context_menu_section {
+  display: flex;
+  align-items: center;
 }
 
 @media screen and (min-width: 992px) {
@@ -51,21 +89,20 @@
   #context-menu {
     position: fixed;
     z-index: 10000;
-    width: 150px;
     background: #1b1a1a;
   }
 
-  #context-menu .item {
+  #context-menu .context_menu_section {
     padding: 8px 10px;
     font-size: 15px;
     color: #eee;
   }
 
-  #context-menu .item:hover {
+  #context-menu .context_menu_section:hover {
     background: #555;
   }
 
-  #context-menu .item .icon_menu {
+  #context-menu .context_menu_section .icon_menu {
     display: inline-block;
     margin-right: 5px;
   }
