+++ src/Pages/Assets/PersonnalAssets.tsx
@@ -1,13 +1,13 @@
-import FileUpload from '../../Components/FileUPload/FileUpload'
-import Sidebar from '../../Components/Sidebar/Sidebar'
+import Assets from '../../Components/Sidebar/Sidebar'
+import Sidebar from '../../Components/PersonnalFolders/Home'
 
 import './PersonnalAssets.scss'
 
 const PersonnalAssets: React.FC = () => {
   return (
     <div style={{ display: 'flex' }}>
+      <Assets />
       <Sidebar />
-      <FileUpload />
     </div>
   )
 }
