+++ src/Routes/Routes.tsx
@@ -2,7 +2,7 @@ import { Router, Switch, Route } from 'react-router-dom'
 import { createBrowserHistory } from 'history'
 import UserProvider from '../Components/Context/UserContext'
 
-import AssetDetails from '../Components/Assets/AssetDetails'
+import AssetDetails from '../Components/PersonnalFolders/AssetDetails'
 
 import PersonnalAssets from '../Pages/Assets/PersonnalAssets'
 import AddNewUser from '../Pages/AddNewUser/AddNewUser'
@@ -14,7 +14,7 @@ import NewPassword from '../Pages/Password/NewPassword'
 
 import Home from '../Pages/Home/Home'
 import Login from '../Pages/Login/Login'
-import Assets from '../Components/Assets/Assets'
+import Assets from '../Components/PersonnalFolders/Home'
 
 const customHistory = createBrowserHistory()
 
@@ -33,7 +33,7 @@ const Routes: React.FC = () => {
               path="/password_management/:token/:id"
               component={NewPassword}
             />
-            <Route exact path="/assets" component={Assets} />
+            <Route exact path="/assets" component={PersonnalAssets} />
             <Route exact path="/register-new-user" component={AddNewUser} />
             <Route exact path="/assets/:id" component={AssetDetails} />
             <Route exact path="/" component={Home} />
